+++ tiskitpy/compliance/__init__.py
@@ -1,4 +1,4 @@
-from .compliance_noise import ComplianceNoise
+from .seafloor_synthetic import SeafloorSynthetic
 from .compliance import Compliance
 from .psd_vals import PSDVals
 from .earth_model import EarthModel1D